Render history page with React
