+++ electron-app/src/renderer/locales/ko.json
@@ -24,9 +24,9 @@
         "output_template_help": "기본 템플릿: %(title)s [%(id)s].%(ext)s"
       },
       "server": {
-        "title": "서버 구성",
+        "title": "서버 설정",
         "port": "서버 포트",
-        "port_help": "서버용 포트 (1024-65535)"
+        "port_help": "HTTP API와 실시간 업데이트에 사용하는 포트 (1024-65535)"
       },
       "application": {
         "title": "애플리케이션 동작",
