+++ server/package.json
@@ -2,16 +2,27 @@
   "name": "ducktracker-server",
   "version": "1.0.0",
   "description": "Lightweight download coordination server for DuckTracker",
-  "main": "src/index.js",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "scripts": {
-    "start": "node src/index.js",
-    "dev": "node --watch src/index.js"
+    "build": "tsc",
+    "start": "node dist/index.js",
+    "dev": "ts-node --transpile-only src/index.ts"
   },
   "keywords": ["youtube", "downloader", "ducktracker"],
   "author": "",
   "license": "MIT",
   "type": "commonjs",
   "dependencies": {
-    "better-sqlite3": "^9.4.1"
+    "better-sqlite3": "^9.4.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1"
+  },
+  "devDependencies": {
+    "@types/node": "^22.10.1",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.7",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.6.3"
   }
 }
